Haxe 3.4 now prints "pack.a.ModuleName.ClassName"
rather than just "pack.a.ModuleName", which it used to do if MoudleName and ClassName were the same.

This changes the naming of our dependencies.  It's technically a breaking change but I don't think most people would notice.  It's only noticeable for us because our tests use the manual naming so that there are less macros (and less moving parts) happening in each test, so they are more specific.