Add new shorter mapping syntax
This allows you to use a plain local variable name ("ident" in macro terminology) and it will create a mapping with the same name and type as that variable, for it's value.

It also allows you to use any other expression that returns a known type and it will create a wildcard mapping for that type.