Make sure we always use `String` instead of `String.String` in module IDs
A tink_macro change caused IDs to change from printing the full module and type,
like `pack.MyClass.MyClass` to just printing the module if the type was the same,
so `pack.MyClass`. I've made this behaviour standard now, even on older versions
of tink_macro.

This does cause many IDs to change, so technically this is a breaking change, but
you will probably only notice it if you use `injector.getFromId('String.String')`
etc.

The tests have been updated to pass.